Integrate the kadena-graph service into the devnet (#26)
This PR integrates the [@kadena/graph](https://github.com/kadena-community/kadena.js/tree/main/packages/apps/graph) GraphQL server for Chainweb into the devnet.

This initial integration of the service isn't optimized for the Nix closure size, so it adds ~400MB to the uncompressed docker image. That's why we're exposing it through a new variant of the devnet called `graph`. When we optimize its closure in the future we will probably include it in the default image.

* PoC graph integration

* Accept list in chainweb.extra-migrations-folders

* Allow nix store path references in landing page

* Flesh out the graph module

* Add version info to graph service section

* Move graph to the kadena-nix repo

* Move the graph service into a new container variant

* Add revLink to the graph package

* Improve the Graph section of the landing page

* Move the graph's PORT variable into process-compose.environment

* Enable prisma logging in graph

* Replace "with" with "let inherit" for clarity

* Add trailing newline to graph.nix