Improve the mining behavior (#28)
This PR improves the mining behavior of the sandbox (the devnet container) by 
* Switching the `chainweb-mining-client` to always be in the `on-demand` mode
* Adding a new `mining-trigger` service that triggers mining:
   * 2 blocks on all chains immediately at start up (to avoid transaction TTL issues) 
   * Periodically
   * In response to new transactions, by intercepting the incoming `/send` requests to the http server (the `8080` port by default)

The default behavior of this `mining-trigger` service is a minor optimization over the previous mining behavior of constantly mining blocks every 5 seconds. Now, by default, blocks are mined every 30 seconds when the sandbox is idle (i.e. no incoming transactions) and whenever transactions arrive, blocks are minted every 5 seconds until the transactions are confirmed by 5 blocks. The purpose of this default behavior is to avoid isolating the sandbox users from the mining periods entirely so that we don't encourage building dApps with user experiences that don't take the mining and confirmation periods into account.

However, it is possible to customize this mining behavior by running the sandbox container by passing it some environment variables. For example, starting the container with `-e MINING_BATCH_PERIOD=0.05 -e MINING_CONFIRMATION_PERIOD=1` will cause it to mint transactions immediately and then mint an additional confirmation blocks every second.

Note that I've published the `kadena/devnet:experimental-smart-mining.graph` image from this branch in order to facilitate testing this PR.

This PR also improves the landing page to document these environment variables:

#### MINING_BATCH_PERIOD
Wait for this period, in seconds, after receiving a transaction and then mine blocks
on the chain where the transaction was received. This period is used to batch transactions
and avoid mining a block for each transaction. Increasing this period also makes mining
more realistic compared to the public networks
 (default: 5.0)

#### MINING_CONFIRMATION_COUNT
The number of blocks to mine faster after transactions; This makes it quicker to get
a transaction confirmed.
 (default: 5)

#### MINING_CONFIRMATION_PERIOD
The period, in seconds, to wait after minting the confirmation blocks of transactions
 (default: 5.0)

#### MINING_IDLE_PERIOD
The average time, in seconds, it takes to mine blocks and advance the block height
by one while the network is idle (i.e. no incoming transactions)
 (default: 30.0)

#### MINING_DISABLE_IDLE
Disable periodic mining when the network is idle. Note that this is *NOT RECOMMENDED*
for most cases, since in the absence of mining, the node's current time will
lag behind and transactions will not be accepted. Consider increasing MINING_IDLE_PERIOD
instead.
 (default: false)

#### MINING_DISABLE_CONFIRMATION
Disable quick mining for confirming transactions. Note that if you want to mint
the blocks containing transactions only and no further confirmations, don't disable
this option and use MINING_CONFIRMATION_COUNT=1 instead.
 (default: false)

* Add a Haskell service for triggering mining

* Insert the mining-trigger into the /send endpoint

* Streamline the chainweb-node readiness-probe

* Make mining-trigger depend on mining-client

* Implement periodic and transaction-based triggering

* Change the default mining-client to on-demand

* Remove the top-level cabal.project

It's not a good idea in the long term since ad-hoc packages
like the mining-trigger don't have to live in a single project
so that they can diverge in various ways

* Improve CWN the request limiting for crashnet

* Further throttling tweaks

* Allow CORS in mining-trigger

* Remove unused import

* Use simpleCorsResourcePolicy

* Avoid duplicate CORS headers

* Improve the mining trigger with debouncing/batching

* Rename mining-trigger/Main.hs -> MiningTrigger.hs

* Request 2 blocks on all chains at start up

Instead of 5 blocks on chain 0 only, which creates more blocks
unnecessarily

* Request blocks on all pending chains simultaneously

* Make mining-trigger configurable

* Make mining-trigger port configurable

* Speed up the time-to-mint for new transactions

* Extract TriggerState to new module

* Move code around

* Streamline the CWD readiness probe

* Make on-demand + mining-trigger only worker type

And make its default behavior be a hybrid between public chainweb and the previous constan-delay mining-client. It can still be configured to mine fast via environment variables.

* Document the mining-related envVars in landing-page

* Improve the flexible mining documentation

* Minor improvement to Flexible Mining section

* Simplify openDetailsIfHashMatches

* Trigger mining only if /send succeeds

* Disable the /make-blocks endpoint on crashnet

* Don't pushTransaction when demand <= 0

* Allow disabling mining workers

* Check that idleTriggerPeriod is positive

* Warn about disabling idle mining

* Be more flexible with disable args

* Be explicit about the MINING_DISABLE_CONFIRMATION variable

* Simplify option parsing code

* Delay periodic blocks when there's network activity

* Implement cooldown for mining blocks

* Make the log messages consistent with arg naming

* Fix unnecessary confirmation blocks

* Accept request logging type as argument

* Use fast-logger for logging

* Carry the HTTP.Manager in AppCtx

* Add newlines to logs

* Replace listening to port message

* Add timestamps to log lines

* Rename Peridic Trigger to Idle Trigger

* Refactor

* Add milliseconds to time format

* Swith to debug logs in chainweb-data

* Make the time format compatible with other  chainweb services

* Don't block pushTransaction

* Merge chain pending count via max